Crate ihex

The IHEX Library

A Rust library for parsing and generating Intel HEX (or IHEX) objects. This format is commonly used for representing compiled program code and data to be loaded into a microcontroller, flash memory or ROM.

Modules

types

Structs

Reader

ReaderOptions

Enums

ReaderError

Record

WriterError

Functions

checksum
Computes the Intel HEX checksum of data. This is done by summing all the bytes `data and taking the two's complement of the least significant byte of the sum.

create_object_file_representation
Generates an Intel HEX object file representation of the records provided. It is the callers responsibility to ensure that no overlapping data ranges are defined within the object file. In addition, records must have contain 1 EoF record, and it must be the last element in records.