Enum ihex::record::Record

pub enum Record {
    Data {
        offset: u16,
        value: Vec<u8>,
    },
    EndOfFile,
    ExtendedSegmentAddress(u16),
    StartSegmentAddress {
        cs: u16,
        ip: u16,
    },
    ExtendedLinearAddress(u16),
    StartLinearAddress(u32),
}

Variants
Data

Specifies a 16-bit offset address and up to 255 bytes of data. Availability: I8HEX, I16HEX and I32HEX.

Fields of Data

offset: u16

The offset of the data record in memory.

value: Vec<u8>

Up to 255 bytes of data to be written to memory.

EndOfFile

Indicates the end of the object file. Must occur exactly once per file, at the end. Availability: I8HEX, I16HEX and I32HEX.

ExtendedSegmentAddress(u16)

Specifies bits 4-19 of the Segment Base Address (SBA) to address up to 1MiB. Availability: I16HEX.

StartSegmentAddress

Specifies the 20-bit segment address via the CS and IP registers. Availability: I16HEX.

Fields of StartSegmentAddress

cs: u16

Value of the CS register.

ip: u16

Value of the IP register.

ExtendedLinearAddress(u16)

Specifies the upper 16 bits of a 32-bit linear address. The lower 16 bits are derived from the Data record load offset. Availability: I32HEX.

StartLinearAddress(u32)

Specifies the execution start address for the object file. This is the 32-bit linear address for register EIP. Availability: I32HEX.

Methods

impl Record

pub fn from_record_string(string: &str) -> Result<Self, ReaderError>

Constructs a new Record by parsing string.

Examples

use ihex::record::Record;

let record = Record::from_record_string(":");

impl Record

pub fn record_type(&self) -> u8

Returns the record type specifier corresponding to the receiver.

impl Record

pub fn to_string(&self) -> Result<String, WriterError>

Returns the IHEX record representation of the receiver, or an error on failure.