Function ihex::writer::create_object_file_representation

pub fn create_object_file_representation(
    records: &[Record]
) -> Result<String, WriterError>

Generates an Intel HEX object file representation of the records provided. It is the callers responsibility to ensure that no overlapping data ranges are defined within the object file. In addition, records must have contain 1 EoF record, and it must be the last element in records.

Example

use ihex::record::Record;
use ihex::writer;

let records = &[
  Record::Data { offset: 0x0010, value: vec![0x48,0x65,0x6C,0x6C,0x6F] },
  Record::EndOfFile
];

let result = writer::create_object_file_representation(records);